Add AND-gate cable node and fix multi-input door wiring
Doors wired to multiple buttons/sensors tracked "should stay open" using
a single shared slot keyed by the door's own net_id, so whichever button
toggled last silently overwrote every other button's vote (closing one
input closed the door even while another wired input was still active).
Fixed by keying each vote by the wiring source's own net_id instead, so
every button/sensor/gate gets an independent slot (proper OR semantics),
and as a side effect a destroyed button's stale vote now gets cleaned up
automatically by the door's existing liveness check.

Also adds a real AND-gate cable connection node (sdNode.TYPE_SIGNAL_AND_GATE):
it discovers its live inputs (buttons/sensors/other AND gates wired through
plain passthrough nodes only) and passes a signal onward only while all of
them are simultaneously active, chainable for multi-stage logic. The button
press propagation logic (doors/antigravity/turrets/nodes) was extracted out
of sdButton.SetActivated into a shared sdSignalPropagation.js so the new
gate type can drive the exact same downstream behavior a button would.
