+++ game/entities/sdButton.js
@@ -17,13 +17,7 @@ import sdBlock from './sdBlock.js';
 import sdGun from './sdGun.js';
 import sdDoor from './sdDoor.js';
 import sdBaseShieldingUnit from './sdBaseShieldingUnit.js';
-import sdAntigravity from './sdAntigravity.js';
-import sdNode from './sdNode.js';
-import sdTurret from './sdTurret.js';
-import sdSampleBuilder from './sdSampleBuilder.js';
-import sdSteeringWheel from './sdSteeringWheel.js';
-import sdLiquidAbsorber from './sdLiquidAbsorber.js';
-import sdConveyor from './sdConveyor.js';
+import { PropagateSignal } from './sdSignalPropagation.js';
 
 import sdSound from '../sdSound.js';
 
@@ -506,382 +500,9 @@ class sdButton extends sdEntity
 		{
 			this.activated = v;
 			this._update_version++;
-			
-			let nodes_to_switch_off = [];
-			
-			const GetFlipLogic = ( path )=>
-			{
-				let vv = v;
-				
-				for ( let i = 0; i < path.length; i++ )
-				{
-					let node = path[ i ];
-					if ( node.is( sdNode ) )
-					{
-						if ( node.type === sdNode.TYPE_SIGNAL_ONCE )
-						{
-							if ( nodes_to_switch_off.indexOf( node ) === -1 )
-							nodes_to_switch_off.push( node );
-						}
-						else
-						if ( node.type === sdNode.TYPE_SIGNAL_ONCE_OFF )
-						{
-							return undefined;
-						}
-						else
-						if ( node.type === sdNode.TYPE_SIGNAL_FLIPPER )
-						vv = !vv;
-					}
-				}
-		
-				return vv;
-			};
-			
-			let doors = this.FindObjectsInACableNetwork( null, sdDoor, true ); // { entity: sdEntity, path: [] }
-			let entities_with_toggle_enabled = [ 
-				...this.FindObjectsInACableNetwork( null, sdAntigravity, true ), 
-				...this.FindObjectsInACableNetwork( null, sdSampleBuilder, true ), 
-				...this.FindObjectsInACableNetwork( null, sdSteeringWheel, true ), 
-				...sdConveyor.AppendConnectedConveyorsToArray( this.FindObjectsInACableNetwork( null, sdConveyor, true ) ), 
-				...this.FindObjectsInACableNetwork( null, sdLiquidAbsorber, true ) 
-			]; // { entity: sdEntity, path: [] }
-			let nodes = this.FindObjectsInACableNetwork( null, sdNode, true );
-			let turrets = this.FindObjectsInACableNetwork( null, sdTurret, true );
-			
-			// Prevents path building through BSUs, LRTPs etc. Only nodes are allowed
-			const IsPathTraversable = ( path )=>
-			{
-				for ( let i2 = 1; i2 < path.length; i2++ ) // 0 is button/sensor
-				if ( !path[ i2 ].is( sdNode ) )
-				return false;
-				
-				return true;
-			};
-			
-			const MeasureDelayAndDo = ( path, ent, vv, then )=>
-			{
-				let delay = 0;
-				
-				/*if ( !this.activated )
-				{
-					// Skip delays on deactivation? This will allow ping-pong movement for elevator movements when delay is added
-				}
-				else*/
-				for ( let i2 = 0; i2 < path.length; i2++ ) // 0 is button/sensor
-				{
-					let node = path[ i2 ];
-					
-					let i2_copy = i2;
-					//
-					if ( node.is( sdNode ) )
-					if ( node.type === sdNode.TYPE_SIGNAL_DELAYER )
-					{
-						setTimeout( ()=>
-						{
-							
-							let fail = false;
-							for ( let i2 = 0; i2 <= i2_copy; i2++ )
-							if ( path[ i2 ]._is_being_removed )
-							{
-								fail = true;
-								break;
-							}
-							
-							//if ( !node._is_being_removed )
-							if ( !fail )
-							{
-								node.variation = vv ? 1 : 2;
-								node._update_version++;
-								
-								sdSound.PlaySound({ name:'sd_beacon_disarm', x:node.x, y:node.y, volume:0.5, pitch:8 });
-							}
-						}, delay );
-						setTimeout( ()=>
-						{
-							if ( !node._is_being_removed )
-							{
-								node.variation = 0;
-								node._update_version++;
-							}
-						}, delay + node.delay );
 
+			PropagateSignal( this, v );
 
-						delay += node.delay;
-					}
-				}
-				
-				if ( delay === 0 )
-				then();
-				else
-				{
-					setTimeout( ()=>{
-						if ( !ent._is_being_removed )
-						{
-							let fail = false;
-							for ( let i2 = 0; i2 < path.length; i2++ )
-							if ( path[ i2 ]._is_being_removed )
-							{
-								fail = true;
-								break;
-							}
-
-							if ( !fail )
-							then();
-						}
-					}, delay );
-				}
-			};
-			
-			for ( let i = 0; i < nodes.length; i++ )
-			{
-				let node = nodes[ i ].entity;
-				
-				if ( node._is_being_removed )
-				continue;
-			
-				let path = nodes[ i ].path;
-				
-				if ( !IsPathTraversable( path ) )
-				continue;
-				
-				let found_switch_off = false;
-				
-				for ( let i2 = 0; i2 < path.length; i2++ )
-				//for ( let i2 = path.length - 1; i2 >= 0; i2-- )
-				{
-					let earlier_node = path[ i2 ];
-					
-					if ( earlier_node.is( sdNode ) )
-					{
-						if ( earlier_node.type === sdNode.TYPE_SIGNAL_ONCE )
-						{
-							found_switch_off = true;
-
-							if ( nodes_to_switch_off.indexOf( earlier_node ) === -1 )
-							nodes_to_switch_off.push( earlier_node );
-
-							break;
-						}
-						
-						if ( earlier_node.type === sdNode.TYPE_SIGNAL_ONCE_OFF )
-						{
-							found_switch_off = true;
-							break;
-						}
-					}
-				}
-				
-				if ( !found_switch_off )
-				if ( node.type === sdNode.TYPE_SIGNAL_ONCE )
-				{
-					if ( nodes_to_switch_off.indexOf( node ) === -1 )
-					nodes_to_switch_off.push( node );
-				}
-			}
-			
-			for ( let i = 0; i < doors.length; i++ )
-			{
-				let door = doors[ i ].entity;
-				
-				if ( door._is_being_removed )
-				continue;
-			
-				let path = doors[ i ].path;
-				
-				let vv = GetFlipLogic( path );
-				
-				if ( vv === undefined )
-				continue;
-				
-				if ( !IsPathTraversable( path ) )
-				continue;
-			
-				MeasureDelayAndDo( path, door, vv, ()=>
-				{
-					door.open_type = sdDoor.OPEN_TYPE_BUTTON;
-
-					let id = door._entities_within_sensor_area.indexOf( door._net_id ); // Add itself so multiple sensors/buttons could be used
-
-					if ( vv )
-					{
-						if ( id === -1 )
-						{
-							door._entities_within_sensor_area.push( door._net_id );
-							door.Open();
-							
-							door._owner = this._owner;
-						}
-					}
-					else
-					if ( id !== -1 )
-					{
-						door._entities_within_sensor_area.splice( id, 1 );
-						door.opening_tim = 0.000001; // Micro value so sound can play
-						
-						door._owner = this._owner;
-					}
-				});
-			}
-
-			//trace( entities_with_toggle_enabled );
-			for ( let i = 0; i < entities_with_toggle_enabled.length; i++ )
-			{
-				let antigravity = entities_with_toggle_enabled[ i ].entity;
-				
-				if ( antigravity._is_being_removed )
-				continue;
-			
-				let path = entities_with_toggle_enabled[ i ].path;
-				
-				let vv = GetFlipLogic( path );
-				
-				if ( vv === undefined )
-				continue;
-				
-				if ( !IsPathTraversable( path ) )
-				continue;
-			
-				/*for ( let i2 = 0; i2 < path.length; i2++ )
-				if ( path[i2]._net_id === 161235055 )
-				{
-					trace('visited 161235055 node');
-				}*/
-				
-				MeasureDelayAndDo( path, antigravity, vv, ()=>
-				{
-					if ( typeof antigravity._update_version !== 'undefined' )
-					antigravity._update_version++;
-
-					if ( typeof antigravity._toggle_source_current !== 'undefined' )
-					{
-						if ( antigravity._toggle_source_current && !antigravity._toggle_source_current._is_being_removed )
-						antigravity._toggle_source_current.SetActivated( false );
-						
-						antigravity._toggle_source_current = vv ? this : null;
-					}
-					
-					if ( typeof antigravity._toggle_direction_current !== 'undefined' )
-					{
-						antigravity._toggle_direction_current = null;
-						for ( let i = path.length - 1; i >= 0; i-- )
-						{
-							let e = path[ i ]
-							if ( ( e.is( sdButton ) && ( e.kind >= sdButton.BUTTON_KIND_TAP_UP && e.kind <= sdButton.BUTTON_KIND_TAP_RIGHT ) ) ||
-								 ( e.is( sdNode ) && ( e.type === sdNode.TYPE_SIGNAL_TURRET_ENABLER ) ))
-							{
-								antigravity._toggle_direction_current = e;
-								break;
-							}
-						}
-					}
-					
-
-					antigravity.toggle_enabled = vv;
-					antigravity.onToggleEnabledChange();
-					
-					if ( vv )
-					{
-						antigravity.SetHiberState( sdEntity.HIBERSTATE_ACTIVE );
-						
-						if ( typeof antigravity._owner !== 'undefined' )
-						antigravity._owner = this._owner;
-						
-						if ( typeof antigravity.owner !== 'undefined' )
-						antigravity.owner = this.owner;
-					}
-				});
-			}
-			for ( let i = 0; i < turrets.length; i++ )
-			{
-				let turret = turrets[ i ].entity;
-				
-				if ( turret._is_being_removed )
-				continue;
-			
-				let path = turrets[ i ].path;
-				
-				let vv = GetFlipLogic( path );
-				
-				if ( vv === undefined )
-				continue;
-				
-				if ( !IsPathTraversable( path ) )
-				continue;
-				
-				
-				MeasureDelayAndDo( path, turret, vv, ()=>
-				{
-					turret._update_version++;
-
-					if ( vv )
-					{
-						// Find last node and use its' direction
-						for ( let i = path.length - 1; i >= 0; i-- )
-						{
-							let node = path[ i ];
-							if ( node.is( sdNode ) )
-                            {
-                                if ( node.type === sdNode.TYPE_SIGNAL_TURRET_ENABLER )
-                                {
-                                    turret.auto_attack = node.variation;
-                                    turret._auto_attack_reference = node;
-                                    break;
-                                }
-                                /*if ( node.type === sdNode.TYPE_SIGNAL_FLIPPER )
-                                {
-                                    if ( turret.kind === sdTurret.KIND_AUTO_CABLE ) // Disconnect all cables that were created by it in case of signal flipping node
-                                    {
-                                        for ( const cable of turret._built_cables )
-                                        {
-                                            if ( cable && !cable._is_being_removed )
-                                            cable.remove();
-                                        }
-                                        turret._built_cables = [];
-                                       break;
-                                    }
-                                }*/
-                            }
-                        }
-
-						turret.SetHiberState( sdEntity.HIBERSTATE_ACTIVE );
-						
-                        if ( turret.kind !== sdTurret.KIND_AUTO_CABLE )
-						turret._owner = this._owner;
-					}
-					else
-					{  
-                        for ( let i = path.length - 1; i >= 0; i-- )
-                        {
-                            let node = path[ i ];
-                            if ( node.is( sdNode ) )
-
-                            if ( turret.kind === sdTurret.KIND_AUTO_CABLE && node.type === sdNode.TYPE_SIGNAL_FLIPPER )  // Disconnect all cables that were created by it in case of signal flipping node
-                            {
-                                for ( const cable of turret._built_cables )
-                                {
-                                    if ( cable && !cable._is_being_removed )
-                                    cable.remove();
-                                }
-                                turret._built_cables = [];
-                            }
-                            else
-                            turret.auto_attack = -1;
-                            turret._auto_attack_reference = null;
-                        }
-					}
-				});
-			}
-			
-			
-			// Then burn all the one-time nodes
-			for ( let i = 0; i < nodes_to_switch_off.length; i++ )
-			{
-				let node = nodes_to_switch_off[ i ];
-				
-				node.type = sdNode.TYPE_SIGNAL_ONCE_OFF;
-				node._update_version++;
-			}
-			
 			if ( sdWorld.time > this._last_sound + 150 )
 			{
 				this._last_sound = sdWorld.time;
